sfc-ostack/dev_requirements.txt
addict==2.1.1
appdirs==1.4.3
asn1crypto==0.23.0
astroid==1.5.3
autopep8==1.3.3
Babel==2.5.0
bcrypt==3.1.3
certifi==2017.7.27.1
cffi==1.11.0
chardet==3.0.4
cliff==2.8.0
cmd2==0.7.7
cryptography==2.0.3
debtcollector==1.17.0
decorator==4.1.2
deprecation==1.0.1
Fabric3==1.13.1.post1
flake8==3.4.1
gitchangelog==3.0.3
idna==2.6
ipdb==0.10.3
ipython==6.1.0
ipython-genutils==0.2.0
iso8601==0.1.12
isort==4.2.15
jedi==0.10.2
jsonpatch==1.16
jsonpointer==1.10
keystoneauth1==3.1.0
lazy-object-proxy==1.3.1
mccabe==0.6.1
monotonic==1.3
msgpack-python==0.4.8
netaddr==0.7.19
netifaces==0.10.6
openstacksdk==0.9.18
os-client-config==1.28.0
osc-lib==1.7.0
oslo.config==4.11.0
oslo.i18n==3.17.0
oslo.serialization==2.20.0
oslo.utils==3.28.0
paramiko==2.3.1
pbr==3.1.1
pexpect==4.2.1
pickleshare==0.7.4
positional==1.2.1
prettytable==0.7.2
prompt-toolkit==1.0.15
ptyprocess==0.5.2
py==1.4.34
pyasn1==0.3.6
pycodestyle==2.3.1
pycparser==2.18
pyflakes==1.5.0
Pygments==2.2.0
pylint==1.7.4
PyNaCl==1.1.2
pyparsing==2.2.0
pyperclip==1.5.27
pytest==3.2.3
python-heatclient==1.11.0
python-keystoneclient==3.13.0
python-neutronclient==6.5.0
python-novaclient==9.1.1
python-openstacksdk==0.5.2
python-swiftclient==3.4.0
pytz==2017.2
PyYAML==3.12
requests==2.18.4
requestsexceptions==1.3.0
rfc3986==1.1.0
simplegeneric==0.8.1
simplejson==3.11.1
six==1.11.0
stevedore==1.25.0
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
wrapt==1.10.11
